scheduler/round_robin: install context on all cpus makes it possible to access other cpus' contexts by using a linked list

+++ kernel/src/arch/x86_64/context/context.h
@@ -7,7 +7,7 @@ typedef struct
     bool type; // false for cpu, true for scheduler
 
     // cpu identification
-    uint16_t cpu_id;
+    uint64_t cpu_id;
     bool is_bsp;
 
     // internal tables
